Addresses PR comments removed  SchemaValidationFunctions.kt and merged functions into SchemaValidation.kt Updated RabbitMQProcessor() and ServiceBusProcessor() to use functions from SchemaValidation.kt Simplified loadSchemaFile() by removing the schemaDirectoryPath parameter, since the schema is retrieved using javaClass.classLoader.getResource() and folder name is constant schema.
